backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pypdf
sentence-transformers
faiss-cpu
python-multipart